Enum ethstore::dir::SetKeyError

pub enum SetKeyError {
    Fatal(Error),
    NonFatalOld(Error),
    NonFatalNew(Error),
}

VaultKeyDirectory::set_key error

Variants
Fatal(Error)

Error is fatal and directory is probably in inconsistent state

NonFatalOld(Error)

Error is non fatal, directory is reverted to pre-operation state

NonFatalNew(Error)

Error is non fatal, directory is consistent with new key