Struct ethstore::EthStore

pub struct EthStore { /* fields omitted */ }

Accounts store.

Methods

impl EthStore

pub fn open(directory: Box<KeyDirectory>) -> Result<Self, Error>

Open a new accounts store with given key directory backend.

pub fn open_with_iterations(
directory: Box<KeyDirectory>,
iterations: u32
) -> Result<Self, Error>

Open a new account store with given key directory backend and custom number of iterations.

pub fn set_refresh_time(&self, time: Duration)

Modify account refresh timeout - how often they are re-read from KeyDirectory.

Setting this to low values (or 0) will cause new accounts to be picked up quickly, although it may induce heavy disk reads and is not recommended if you manage many keys (say over 10k).

By default refreshing is disabled, so only accounts created using this instance of EthStore are taken into account.