Crate ethstore

Ethereum key-management.

Modules

accounts_dir

Accounts Directory

ethkey

ethkey reexport to make documentation look pretty.

Structs

Crypto

Encrypted data

EthMultiStore

Similar to EthStore but may store many accounts (with different passwords) for the same Address

EthStore

Accounts store.

IndexDerivation

Node in hierarchical derivation.

KeyFile

Public opaque type representing serializable KeyFile.

OpaqueSecret

An opaque wrapper for secret.

PresaleWallet

Pre-sale wallet.

SafeAccount

Account representation.

StoreAccountRef

Stored account reference

Enums

Derivation

Derivation scheme for keys

Error

Account-related errors.

SecretVaultRef

Key directory reference

Traits

SecretStore

Secret Store API

SimpleSecretStore

Simple Secret Store API

Functions

import_account

Import an account from a file.

import_accounts

Import all accounts from one directory to the other.

random_phrase

Generate a string which is a random phrase of a number of lowercase words.

random_string

Generate a random string of given length.