Struct ethstore::SafeAccount

pub struct SafeAccount {
    pub id: [u8; 16],
    pub version: Version,
    pub address: Address,
    pub crypto: Crypto,
    pub filename: Option<String>,
    pub name: String,
    pub meta: String,
}

Fields
id: [u8; 16] version: Version address: Address crypto: Crypto filename: Option<String> name: String meta: String
Methods

impl SafeAccount

fn create(keypair: &KeyPair,
id: [u8; 16],
password: &str,
iterations: u32,
name: String,
meta: String)
-> Self

fn from_file(json: KeyFile, filename: Option<String>) -> Self

Create a new SafeAccount from the given json; if it was read from a file, the filename should be Some name. If it is as yet anonymous, then it can be left None.

fn sign(&self, password: &str, message: &Message) -> Result<Signature, Error>

fn decrypt(&self,
password: &str,
shared_mac: &[u8],
message: &[u8])
-> Result<Vec<u8>, Error>

fn public(&self, password: &str) -> Result<Public, Error>

fn change_password(&self,
old_password: &str,
new_password: &str,
iterations: u32)
-> Result<Self, Error>

fn check_password(&self, password: &str) -> bool